LCORE-641: split unit tests for HTTP requests models
